Feat: Move tags editor to sidebar
